+++ src/StkTable/StkTable.vue
@@ -1037,22 +1037,25 @@ function onTrMouseOver(_e: MouseEvent, row: DT) {
 
 /**
  * 选中一行，
- * @param {string} rowKey selected rowKey, undefined to unselect
+ * @param {string} rowKeyOrRow selected rowKey, undefined to unselect
  * @param {boolean} option.silent if emit current-change. default:false(not emit `current-change`)
  */
-function setCurrentRow(rowKey: string | undefined, option = { silent: false }) {
+function setCurrentRow(rowKeyOrRow: string | undefined | DT, option = { silent: false }) {
     if (!dataSourceCopy.value.length) return;
-    if (rowKey === void 0) {
+    if (rowKeyOrRow === void 0) {
         currentRow.value = void 0;
         currentRowKey.value = void 0;
-    } else {
-        const row = dataSourceCopy.value.find(it => rowKeyGen(it) === rowKey);
+    } else if (typeof rowKeyOrRow === 'string') {
+        const row = dataSourceCopy.value.find(it => rowKeyGen(it) === rowKeyOrRow);
         if (!row) {
-            console.warn('setCurrentRow failed.rowKey:', rowKey);
+            console.warn('setCurrentRow failed.rowKey:', rowKeyOrRow);
             return;
         }
         currentRow.value = row;
-        currentRowKey.value = rowKey;
+        currentRowKey.value = rowKeyOrRow;
+    } else {
+        currentRow.value = rowKeyOrRow;
+        currentRowKey.value = rowKeyGen(rowKeyOrRow);
     }
     if (!option.silent) {
         emits('current-change', /** no Event */ null, currentRow.value, { select: true });